requirements.txt
# Framework web
flask>=2.0.0

# Variables de entorno
python-dotenv>=0.19.0

# Análisis de datos y machine learning
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.2.0

# Persistencia de modelos
joblib>=1.2.0

# Conexión a base de datos PostgreSQL
psycopg2-binary>=2.9.0

# Herramientas adicionales
python-dateutil>=2.8.0
pytz>=2022.0

# Para debugging y desarrollo
requests>=2.28.0

# Librerías estándar de Python (incluidas por defecto pero listadas por claridad)
# subprocess - incluido en Python estándar
# os - incluido en Python estándar  
# datetime - incluido en Python estándar
# traceback - incluido en Python estándar